requirements.txt
amqp==5.3.1
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.4
cachetools==5.5.0
celery==5.5.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==3.1.0
Django==4.2
django-allauth==64.1.0
django-appconf==1.0.6
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-compressor==4.5.1
django-easy-audit==1.3.7
django-editorjs-fields==0.2.7
django-environ==0.11.2
django-jazzmin==3.0.0
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
et-xmlfile==1.1.0
feedparser==6.0.11
freetype-py==2.5.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.64.0
gunicorn==21.2.0
honcho==2.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
jwt==1.3.1
kombu==5.5.4
lxml==6.0.2
Naked==0.1.32

oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.1.4
pillow==10.4.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.29.0
pycparser==2.22
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyparsing==3.1.4
pypdf==6.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
razorpay==2.0.0
rcssmin==1.1.2
redis==6.4.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rjsmin==1.2.2
rlPyCairo==0.4.0
rsa==4.9
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.8.3
sqlparse==0.5.1
svglib==1.6.0
tinycss2==1.5.1
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.1.1
uritools==6.0.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
xlrd==2.0.1
yapf==0.43.0
